+++ yosai/core/__init__.py
@@ -32,6 +32,7 @@ from .exceptions import (
     AdditionalAuthenticationRequired,
     AuthenticationException,
     AuthorizationException,
+    ConsumedTOTPToken,
     ExpiredSessionException,
     IdleExpiredSessionException,
     IncorrectCredentialsException,