+++ yosai/core/mgt/mgt.py
@@ -450,7 +450,7 @@ class NativeSecurityManager(mgt_abcs.SecurityManager):
         :param role_s: 1..N role identifiers (strings)
         :type role_s:  Set of Strings
 
-        :returns: a frozenset of tuple(s), containing the role and a Boolean
+        :returns: a set of tuple(s), containing the role and a Boolean
                   indicating whether the user is a member of the Role
         """
         return self.authorizer.has_role(identifiers, role_s)