+++ CHANGES.md
@@ -35,6 +35,7 @@ v0.3
   instantiating Permission instances on an as-needed basis
 - permission-based authorization checks no longer supports a list of Permission instances
   as an argument but rather only list of string-typed permissions
+- successful TOTP authentication raises a ConsumedTOTPToken exception
 
 Yosai uses [GitHub's Releases feature](https://github.com/blog/1547-release-your-software) for its changelogs.